requirements.txt
pandas==1.3.3
psycopg2-binary==2.9.1
pyyaml==5.4.1
